https://github.com/ITISFoundation/osparc-simcore/issues/3486 refactoring
