+++ packages/models-library/src/models_library/services_enums.py
@@ -40,3 +40,24 @@ class ServiceState(Enum):
             ServiceState.STOPPING: 5,
             ServiceState.COMPLETE: 6,
         }
+
+
+class ServiceType(str, Enum):
+    COMPUTATIONAL = "computational"
+    DYNAMIC = "dynamic"
+    FRONTEND = "frontend"
+    BACKEND = "backend"
+
+
+# TODO: create a flags enum that accounts for every column
+#
+# | service name    | defininition | implementation | runs                    | ``ServiceType``               |                 |
+# | --------------- | ------------ | -------------- | ----------------------- | ----------------------------- | --------------- |
+# | ``file-picker`` | BE           | FE             | FE                      | ``ServiceType.FRONTEND``      | function        |
+# | ``isolve``      | DI-labels    | DI             | Dask-BE (own container) | ``ServiceType.COMPUTATIONAL`` | container       |
+# | ``jupyter-*``   | DI-labels    | DI             | DySC-BE (own container) | ``ServiceType.DYNAMIC``       | container       |
+# | ``iterator-*``  | BE           | BE             | BE    (webserver)       | ``ServiceType.BACKEND``       | function        |
+# | ``pyfun-*``     | BE           | BE             | Dask-BE  (dask-sidecar) | ``ServiceType.COMPUTATIONAL`` | function        |
+#
+#
+# where FE (front-end), DI (docker image), Dask/DySC (dask/dynamic sidecar), BE (backend).
