+++ packages/models-library/src/models_library/services_io.py
@@ -1,0 +1,249 @@
+from typing import Any, ClassVar
+
+from pydantic import (
+    BaseModel,
+    Extra,
+    Field,
+    StrictBool,
+    StrictFloat,
+    StrictInt,
+    validator,
+)
+
+from .services_constants import PROPERTY_TYPE_RE
+from .services_constrained_types import FileName, ServicePortKey
+from .services_ui import Widget
+from .utils.json_schema import (
+    InvalidJsonSchema,
+    any_ref_key,
+    jsonschema_validate_schema,
+)
+
+
+class BaseServiceIOModel(BaseModel):
+    """
+    Base class for service input/outputs
+    """
+
+    ## management
+
+    ### human readable descriptors
+    display_order: float | None = Field(
+        None,
+        alias="displayOrder",
+        deprecated=True,
+        description="DEPRECATED: new display order is taken from the item position. This will be removed.",
+    )
+
+    label: str = Field(..., description="short name for the property", example="Age")
+    description: str = Field(
+        ...,
+        description="description of the property",
+        example="Age in seconds since 1970",
+    )
+
+    # mathematical and physics descriptors
+    property_type: str = Field(
+        ...,
+        alias="type",
+        description="data type expected on this input glob matching for data type is allowed",
+        examples=[
+            "number",
+            "boolean",
+            "data:*/*",
+            "data:text/*",
+            "data:[image/jpeg,image/png]",
+            "data:application/json",
+            "data:application/json;schema=https://my-schema/not/really/schema.json",
+            "data:application/vnd.ms-excel",
+            "data:text/plain",
+            "data:application/hdf5",
+            "data:application/edu.ucdavis@ceclancy.xyz",
+        ],
+        regex=PROPERTY_TYPE_RE,
+    )
+
+    content_schema: dict[str, Any] | None = Field(
+        None,
+        description="jsonschema of this input/output. Required when type='ref_contentSchema'",
+        alias="contentSchema",
+    )
+
+    # value
+    file_to_key_map: dict[FileName, ServicePortKey] | None = Field(
+        None,
+        alias="fileToKeyMap",
+        description="Place the data associated with the named keys in files",
+        examples=[{"dir/input1.txt": "key_1", "dir33/input2.txt": "key2"}],
+    )
+
+    # TODO: should deprecate since content_schema include units
+    unit: str | None = Field(
+        None,
+        description="Units, when it refers to a physical quantity",
+    )
+
+    class Config:
+        extra = Extra.forbid
+
+    @validator("content_schema")
+    @classmethod
+    def check_type_is_set_to_schema(cls, v, values):
+        if v is not None and (ptype := values["property_type"]) != "ref_contentSchema":
+            msg = f"content_schema is defined but set the wrong type.Expected type=ref_contentSchema but got ={ptype}."
+            raise ValueError(msg)
+        return v
+
+    @validator("content_schema")
+    @classmethod
+    def check_valid_json_schema(cls, v):
+        if v is not None:
+            try:
+                jsonschema_validate_schema(schema=v)
+
+                if any_ref_key(v):
+                    # SEE https://github.com/ITISFoundation/osparc-simcore/issues/3030
+                    msg = "Schemas with $ref are still not supported"
+                    raise ValueError(msg)
+
+            except InvalidJsonSchema as err:
+                failed_path = "->".join(map(str, err.path))
+                msg = f"Invalid json-schema at {failed_path}: {err.message}"
+                raise ValueError(msg) from err
+        return v
+
+    @classmethod
+    def _from_json_schema_base_implementation(
+        cls, port_schema: dict[str, Any]
+    ) -> dict[str, Any]:
+        description = port_schema.pop("description", port_schema["title"])
+        return {
+            "label": port_schema["title"],
+            "description": description,
+            "type": "ref_contentSchema",
+            "contentSchema": port_schema,
+        }
+
+
+class ServiceInput(BaseServiceIOModel):
+    """
+    Metadata on a service input port
+    """
+
+    # TODO: should deprecate since content_schema include defaults as well
+    default_value: StrictBool | StrictInt | StrictFloat | str | None = Field(
+        None, alias="defaultValue", examples=["Dog", True]
+    )
+
+    widget: Widget | None = Field(
+        None,
+        description="custom widget to use instead of the default one determined from the data-type",
+    )
+
+    class Config(BaseServiceIOModel.Config):
+        schema_extra: ClassVar[dict[str, Any]] = {
+            "examples": [
+                # file-wo-widget:
+                {
+                    "displayOrder": 1,
+                    "label": "Input files - file-wo-widget",
+                    "description": "Files downloaded from service connected at the input",
+                    "type": "data:*/*",
+                },
+                # v2
+                {
+                    "displayOrder": 2,
+                    "label": "Sleep Time - v2",
+                    "description": "Time to wait before completion",
+                    "type": "number",
+                    "defaultValue": 0,
+                    "unit": "second",
+                    "widget": {"type": "TextArea", "details": {"minHeight": 3}},
+                },
+                # latest:
+                {
+                    "label": "Sleep Time - latest",
+                    "description": "Time to wait before completion",
+                    "type": "number",
+                    "defaultValue": 0,
+                    "unit": "second",
+                    "widget": {"type": "TextArea", "details": {"minHeight": 3}},
+                },
+                {
+                    "label": "array_numbers",
+                    "description": "Some array of numbers",
+                    "type": "ref_contentSchema",
+                    "contentSchema": {
+                        "title": "list[number]",
+                        "type": "array",
+                        "items": {"type": "number"},
+                    },
+                },
+                {
+                    "label": "my_object",
+                    "description": "Some object",
+                    "type": "ref_contentSchema",
+                    "contentSchema": {
+                        "title": "an object named A",
+                        "type": "object",
+                        "properties": {
+                            "i": {"title": "Int", "type": "integer", "default": 3},
+                            "b": {"title": "Bool", "type": "boolean"},
+                            "s": {"title": "Str", "type": "string"},
+                        },
+                        "required": ["b", "s"],
+                    },
+                },
+            ],
+        }
+
+    @classmethod
+    def from_json_schema(cls, port_schema: dict[str, Any]) -> "ServiceInput":
+        """Creates input port model from a json-schema"""
+        data = cls._from_json_schema_base_implementation(port_schema)
+        return cls.parse_obj(data)
+
+
+class ServiceOutput(BaseServiceIOModel):
+    widget: Widget | None = Field(
+        None,
+        description="custom widget to use instead of the default one determined from the data-type",
+        deprecated=True,
+    )
+
+    class Config(BaseServiceIOModel.Config):
+        schema_extra: ClassVar[dict[str, Any]] = {
+            "examples": [
+                {
+                    "displayOrder": 2,
+                    "label": "Time Slept",
+                    "description": "Time the service waited before completion",
+                    "type": "number",
+                },
+                {
+                    "displayOrder": 2,
+                    "label": "Time Slept - units",
+                    "description": "Time the service waited before completion",
+                    "type": "number",
+                    "unit": "second",
+                },
+                {
+                    "label": "Time Slept - w/o displayorder",
+                    "description": "Time the service waited before completion",
+                    "type": "number",
+                    "unit": "second",
+                },
+                {
+                    "label": "Output file 1",
+                    "displayOrder": 4.0,
+                    "description": "Output file uploaded from the outputs folder",
+                    "type": "data:*/*",
+                },
+            ]
+        }
+
+    @classmethod
+    def from_json_schema(cls, port_schema: dict[str, Any]) -> "ServiceOutput":
+        """Creates output port model from a json-schema"""
+        data = cls._from_json_schema_base_implementation(port_schema)
+        return cls.parse_obj(data)
